factored out palette helper additions temporarily to ease merging
