+++ MaterialDesignThemes.Wpf/MaterialDesignThemes.Wpf.csproj
@@ -360,6 +360,7 @@
     <Compile Include="DialogHost.cs" />
     <Compile Include="DpiHelper.cs" />
     <Compile Include="DrawerHost.cs" />
+    <Compile Include="EnhancedPaletteHelper.cs" />
     <Compile Include="Extensions.cs" />
     <Compile Include="Flipper.cs" />
     <Compile Include="HexColorItem.cs" />
