+++ MaterialDesignThemes.Wpf/Themes/MaterialDesignTheme.HexColorPicker.xaml
@@ -15,7 +15,6 @@
             <Setter.Value>
                 <ControlTemplate TargetType="{x:Type wpf:HexColorItem}">
                     <Border Background="{TemplateBinding ColorBrush}">
-                        
                     </Border>
                 </ControlTemplate>
             </Setter.Value>
